Refactor: improve code efficiency
